#300: Moved skill_class.php to the sub-folder "includes"
 Sorting things around.
