+++ src/php/includes/skill_class.php
